feat: Update next.config.mjs to use new API endpoint for /api/mp/:path*
The next.config.mjs file has been updated to use the new API endpoint http://api.notion-nice.com/mp/:path* for the /api/mp/:path* route. This change ensures that the correct API is called when accessing the /api/mp/:path* route, improving the functionality and reliability of the application.