feat: Add pay route for processing transactions
The code changes in `route.ts` add a new route for processing transactions. This route handles a `POST` request and requires authentication. It sends a request to the `/pay/transactions` endpoint with the user and transaction type data. If the response is successful, it returns the JSON response. Otherwise, it returns an error response. This change improves the functionality of the application by allowing users to process transactions.